requirements.txt
fastapi
sqlalchemy
pyjwt
fastapi-jwt-auth
